Extract CLI helper methods to a module
For readability sake.
